TheDAOHardForkOracle.sol: first version, uncompiled, untested.
Intended use: send a transaction on both tines of the fork. If block
number is after 1920000, fallback can run once. It checks the balance
of WithdrawDAO, and if that is sufficiently big, sets a `forked`
flag.

That can be used by other dapps by accessor `forked`. Contract throws
on any other further invocation.